Add checks for correct sort and limit parameters.
